move dbnode/serialize  to x/serialize
