+++ src/cmd/services/m3coordinator/downsample/id_pool_types.go
@@ -25,7 +25,6 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/m3db/m3/src/dbnode/serialize"
 	"github.com/m3db/m3/src/x/serialize"
 	"github.com/m3db/m3metrics/metric/id"
 	"github.com/m3db/m3x/ident"
@@ -48,7 +47,7 @@ var (
 
 func isRollupID(
 	sortedTagPairs []byte,
-	iteratorPool *xserialize.EncodedTagsIteratorPool,
+	iteratorPool *serialize.EncodedTagsIteratorPool,
 ) bool {
 	iter := iteratorPool.Get()
 	iter.Reset(sortedTagPairs)
@@ -208,7 +207,7 @@ func (p *rollupIDProviderPool) Put(v *rollupIDProvider) {
 
 func resolveEncodedTagsNameTag(
 	id []byte,
-	iterPool *xserialize.EncodedTagsIteratorPool,
+	iterPool *serialize.EncodedTagsIteratorPool,
 	nameTag []byte,
 ) ([]byte, error) {
 	// ID is always the encoded tags for downsampling IDs
