+++ src/cmd/services/m3coordinator/downsample/options.go
@@ -27,7 +27,7 @@ import (
 	"runtime"
 	"time"
 
-	"github.com/m3db/m3/src/dbnode/serialize"
+	"github.com/m3db/m3/src/x/serialize"
 	"github.com/m3db/m3/src/query/storage"
 	"github.com/m3db/m3/src/x/serialize"
 	"github.com/m3db/m3aggregator/aggregator"
@@ -277,7 +277,7 @@ func (o DownsamplerOptions) newAggregator() (agg, error) {
 type aggPools struct {
 	tagEncoderPool          serialize.TagEncoderPool
 	tagDecoderPool          serialize.TagDecoderPool
-	encodedTagsIteratorPool *xserialize.EncodedTagsIteratorPool
+	encodedTagsIteratorPool *serialize.EncodedTagsIteratorPool
 }
 
 func (o DownsamplerOptions) newAggregatorPools() aggPools {
@@ -289,7 +289,7 @@ func (o DownsamplerOptions) newAggregatorPools() aggPools {
 		o.TagDecoderPoolOptions)
 	tagDecoderPool.Init()
 
-	encodedTagsIteratorPool := xserialize.NewEncodedTagsIteratorPool(tagDecoderPool,
+	encodedTagsIteratorPool := serialize.NewEncodedTagsIteratorPool(tagDecoderPool,
 		o.TagDecoderPoolOptions)
 	encodedTagsIteratorPool.Init()
