+++ src/dbnode/client/fetch_tagged_results_index_iterator.go
@@ -21,7 +21,7 @@
 package client
 
 import (
-	"github.com/m3db/m3/src/dbnode/serialize"
+	"github.com/m3db/m3/src/x/serialize"
 	"github.com/m3db/m3x/ident"
 )
