+++ src/dbnode/client/session_pools.go
@@ -22,7 +22,7 @@ package client
 
 import (
 	"github.com/m3db/m3/src/dbnode/encoding"
-	"github.com/m3db/m3/src/dbnode/serialize"
+	"github.com/m3db/m3/src/x/serialize"
 	"github.com/m3db/m3/src/dbnode/x/xpool"
 	"github.com/m3db/m3x/context"
 	"github.com/m3db/m3x/ident"
