+++ src/dbnode/client/session_write_tagged_test.go
@@ -30,7 +30,7 @@ import (
 	"time"
 
 	"github.com/m3db/m3/src/dbnode/generated/thrift/rpc"
-	"github.com/m3db/m3/src/dbnode/serialize"
+	"github.com/m3db/m3/src/x/serialize"
 	"github.com/m3db/m3/src/dbnode/topology"
 	xmetrics "github.com/m3db/m3/src/dbnode/x/metrics"
 	xm3test "github.com/m3db/m3/src/x/test"
