+++ src/dbnode/client/write_state.go
@@ -24,7 +24,7 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/m3db/m3/src/dbnode/serialize"
+	"github.com/m3db/m3/src/x/serialize"
 	"github.com/m3db/m3/src/dbnode/topology"
 	"github.com/m3db/m3cluster/shard"
 	xerrors "github.com/m3db/m3x/errors"
