+++ src/dbnode/network/server/tchannelthrift/options.go
@@ -21,7 +21,7 @@
 package tchannelthrift
 
 import (
-	"github.com/m3db/m3/src/dbnode/serialize"
+	"github.com/m3db/m3/src/x/serialize"
 	"github.com/m3db/m3x/instrument"
 	"github.com/m3db/m3x/pool"
 )
