+++ src/query/test/mock_pools.go
@@ -24,7 +24,7 @@ import (
 	"sync"
 
 	"github.com/m3db/m3/src/dbnode/encoding"
-	"github.com/m3db/m3/src/dbnode/serialize"
+	"github.com/m3db/m3/src/x/serialize"
 	"github.com/m3db/m3/src/dbnode/x/xpool"
 	"github.com/m3db/m3/src/query/pools"
 	"github.com/m3db/m3x/ident"
