+++ src/x/serialize/decoder_lifecycle_prop_test.go
